split KillUnusedSegmentsTask to processing in smaller chunks (enhanced)
Modify the retrieveUnusedSegments call with a limit, so that arbitrarily
large intervals can be processed incrementally. Add the capability to
use this during the KillUnusedSegments Task.

Previously, the KillUnusedSegmentsTask was improved to improve overlord
locking stability when working with large batches, however the retrieve
stage was not improved. So, while the locking stability situation was
improved, we can now handle bigger segment counts. But, processing too
many segments in a batch was resulting in an OOM due to loading too
many segments into memory. This change rectifies this.